requirements.txt
python-dotenv==1.0.0
python-telegram-bot==20.4
aiosqlite==0.19.0
